Create PR quote comments on correct commits
This works by keeping the rewrite-list during the git rebase that moves
the scaffolding history into the user's repo.

Unfortunately this uses an undocumented part of the rebase
implementation, so it may be fragile. However, I could not find another
way to do it without re-implementing the rebase logic on top of
cherry-pick and moving the commits one at a time.