Construct URLs instead of storing them
Because the port can change on each run, stored URLs can go out-of-date.
Instead we now store the information needed to reconstruct the URL.

The templates for normal Forgejo pages seem to either get the URLs from
the backend or to construct them by string interpolation. The latter is
much simpler for us, since we don't know which links we need at template
evaluation time.